src/requirements.txt
Flask
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
slack-sdk
python-dotenv
requests
